Terra-Tag Example Fixes
Changed Tag examples and added new section for Fallback Styles with
alert message for non-interactive tags. Latest: Refactored examples,
resolved lint errors.

Resolves #1368
