+++ packages/terra-site/src/examples/tag/Index.jsx
@@ -3,7 +3,6 @@ import PropsTable from 'terra-props-table';
 import Markdown from 'terra-markdown';
 import ReadMe from 'terra-tag/docs/README.md';
 import { version } from 'terra-tag/package.json';
-import Alert from 'terra-alert';
 
 // Component Source
 // eslint-disable-next-line import/no-webpack-loader-syntax, import/first, import/no-unresolved, import/extensions
@@ -13,14 +12,6 @@ import TagSrc from '!raw-loader!terra-tag/src/Tag';
 import TagDefault from './TagDefault';
 import TagHref from './HrefTag';
 import TagFallbacks from './TagFallbacks';
-
-// Fallback Styles Message
-const alertTitle = 'Not Reccomended Usage.';
-const alertText = 'These styles are provided for when this compoment is missing recomended elements for best practice use.';
-
-const TagHeadingMarginStyle = {
-  marginTop: '200px',
-};
 
 const TagExamples = () => (
   <div>
@@ -33,12 +24,7 @@ const TagExamples = () => (
     <h2 id="href">Tag with href</h2>
     <TagHref />
     <br />
-    <div className="markdown-body" style={TagHeadingMarginStyle}>
-      <h2 id="fallbacks-header">Fallback Styles</h2>
-    </div>
-    <Alert type={Alert.Opts.Types.INFO} title={alertTitle}>{alertText}</Alert>
     <br />
-    <h3 id="fallbacks">Tag with no onClick or href</h3>
     <TagFallbacks />
   </div>
 );
