+++ packages/terra-tag/src/Tag.scss
@@ -40,7 +40,7 @@
   }
 
   .is-interactive {
-    background-color: var(--terra-tag-interactive-background-color);
+    background-color: var(--terra-tag-interactive-background-color, rgba(28, 31, 33, 0.04));
     color: var(--terra-tag-interactive-color, #0059a8);
     cursor: pointer;
     text-decoration: var(--terra-tag-interactive-text-decoration);
